Fixed bug in static_mesh.vs (forgot w = 1 for points). Modified a bunch of states for debugging. Refactored _run out of src/lwe/boot.cc into src/lwe/application.cc. Vertex shader streams are kept for bytecode (TODO: refactor permutations). Meshes now have a local transform (to heirarchy). Replaced vertex declarations with input_layouts (see src/lwe/d3d11_input_layout.cc). Added constant buffer mapping and unmapping (TODO: extend to index and vertex buffers). Fixed pixel format bugs for depth-stencil buffers. Added vec2, vec3, mat3, mat4, and quat math classes. Added lwe_path_find_basename. No longer prepending model name for mesh materials (meshes are treated globally now). Fixed spacing bug with failure and assertion macros. Added lwe_min, lwe_max, and lwe_clamp. Removed global constant buffers from lwe_render_device_dispatch. lwe_viewports can no longer be fractional. Added simple CSV database generation for compiles (useful for Google Docs). Fixed lwe_asset_manager_find_by_hash crash (NULL dereference). Fixed structure alignment issues with shader blobs. Fixed mipmap issues with texture creation. Removed invalid pixel swizzle for texture compiler. Fixed CPUAccessFlags bug with constant buffer (TODO: extend to index and vertex buffer). Implemented basic unit plane demo to test rendering. Fixed pixel format stride bugs. Added lwe_render_stream_draw, lwe_render_stream_set_viewports, and lwe_render_stream_reset.

+++ include/lwe/viewport.h
@@ -28,8 +28,8 @@
 #include <lwe/foundation.h>
 
 typedef struct lwe_viewport_t {
-  float min_x, min_y;
-  float max_x, max_y;
+  uint32_t min_x, min_y;
+  uint32_t max_x, max_y;
 } lwe_viewport_t;
 
 #endif // _LWE_VIEWPORT_H_
